Kiosk Credentialing App/requirements.txt
importlib_metadata==7.1.0
numpy==1.26.4
pycups==2.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-barcode==0.15.1
python-dotenv==1.0.1
segno==1.6.1
zipp==3.18.1
